Extract the service example into its own component
