+++ apps/cookbook/src/app/examples/examples.module.ts
@@ -8,6 +8,7 @@ import {
   KirbyModule,
 } from '@kirbydesign/designsystem';
 
+import { CodeViewerModule } from '../shared/code-viewer/code-viewer.module';
 import { AccordionExampleModule } from './accordion-example/accordion-example.module';
 import { AvatarExampleModule } from './avatar-example/avatar-example.module';
 import { BadgeExampleModule } from './badge-example/badge-example.module';
@@ -38,6 +39,7 @@ import { ToggleButtonExampleModule } from './toggle-button-example/toggle-button
 import { VirtualScrollExampleModule } from './virtual-scroll-example/virtual-scroll-example.module';
 
 const IMPORTS = [
+  CodeViewerModule,
   ExamplesSharedModule,
   FormFieldExamplesModule,
   ItemExampleModule,
