+++ apps/cookbook/src/app/examples/loading-overlay-example/service/loading-overlay-service-example.component.html
@@ -1,0 +1,29 @@
+<button
+  kirby-button
+  attentionLevel="2"
+  size="lg"
+  expand="block"
+  (click)="showFullscreenLoadingOverlay(true)"
+>
+  Show full page loading overlay
+</button>
+<button
+  kirby-button
+  attentionLevel="2"
+  size="lg"
+  expand="block"
+  (click)="showFullscreenLoadingOverlay(false, true)"
+>
+  Show full page loading overlay that hides content
+</button>
+<button
+  kirby-button
+  attentionLevel="2"
+  size="lg"
+  expand="block"
+  (click)="showFullscreenLoadingOverlay(false)"
+>
+  Show full page loading overlay without backdrop
+</button>
+
+<cookbook-code-viewer [ts]="loadingOverlayServiceExample"></cookbook-code-viewer>
