+++ apps/cookbook/src/app/examples/loading-overlay-example/service/loading-overlay-service-example.ts
